Retract SynthesizeExtMethodReceiver mode when going deeper in overloading resolution
The SynthesizeExtMethodReceiver mode is supposed to be turned on only for the
direct application of of a synthesized receiver to the qualifier of an
extension method selection. previously its lifetime was accidentally extended
when overloading resolution looking at subsequent parameter lists because the
first one was not enough to disambiguate.

Fixes #18745
Fixes #18744