Allow inferred parameter types always, when eta-expanding
Rather than drawing an exemption based on fully defined types, report
the right function arity.  This leads eta-expansion to leave off the
type, so then the parameter type is inferred.  In the test case
pos/i18453 this leads to the type var ?A being instantiated to X, rather
than ?A & ?B being constrained against X & Y, which leads to ?A being
instantiated to X & Y - and then failing to find a Box[X & Y] in scope.