simplify: Drop mapping tparams to tvars
Mapping type params back to type vars is dangerous for match types
because it can lead to one use-site's type var being cached as the
reduction of the match type in a match alias defintion.

It looks like mapping type parameters to type vars is only important
when getting the instance type of a type parameter, because the bounds
could make references to other parameters, including as type arguments.
So, as long as we map then, we can drop this mapping from general
simplification.

There is, however, something going on with capture checking, which I
documented, but perhaps it's due to some underlying bug.  The situation
occurs because boxed A *: T does seem to match case x *: _.  So I kept
the fully-defined forcing and normalization case from simplify.