Remove `stdlib213` from community build tests
This test exists for historical reasons. Now we compile the Scala 2
library in the `scala2-library-bootstrapped` project. This one checks
that it compiles (with `-Ycheck:all`) and is used to package the TASTy
files in the `scala2-library-tasty` project. Furthermore, now there is
a `scala/scala` test that checks that the library can be compiled with
Scala 3.

This test is fundamentally broken but we have not noticed it because we
do not use `-Ycheck:all`. The reason is that when we compile `scala.Predef`
we get a conflict with `scala.runtime.stdLibPatches.Predef` when we patch
the symbols. This also surfaced in #18738.

[Cherry-picked 8c6f4c7cade55ff0951cc8e3c7ae59b0d50e9deb]