Fix #18769: Allow HK type args in Java signatures.
Contrary to what an earlier comment said, we do emit HK type
parameters in Java signatures. They are always unbounded and never
the type of values.

However, they can appear as type arguments to other higher-kinded
types. Previously, an assertion error would trigger in that
situation. We relax the assertion to allow this situation and
emit a correct Java signature.

I manually verified that the generated Java signatures are
consistent with what Scala 2 emits for the same code snippet.