Add missing -Yexplicit-nulls for presentation compiler (#18776)
With a new definition of untyped trees, it is now required to have
`-Yexplicit-nulls` flag in modules that use them in order to have proper
type checking. This PR adds the missing flag.

Without the flag, it was possible to first assign untyped trees to typed
trees, and secondly use extension methods for typed trees which can be
seen at `KeywordsCompletions.scala` with usage of
`untpdTree.filterSubtrees`.

It is also blocked by: https://github.com/lampepfl/dotty/issues/18775

I can also make a workaround in the unmanaged module, but it will
require a dependency on nightly version / a new release from metals.