Avoid crashes on missing positions
All positions should be set on these trees and symbols at this point. If
one is missing it is due to a compiler bug. We default the position to 0
if there is a missing position to avoid crashing when calling methods on
Position. To make it possible for us to fix these missing positions, we
warn if there is a missing position when compiling with -Xcheck-macros.