Revert "Fix callTrace if inlined methods"
This reverts commit ec75826f6ff78d0a04d3f93e6ca92bad9308f75d.

The previous encoding of INLINE nodes in TASTy assumed that the callTrace
was a type. We cannot make it a term, or we would not be able to unpickle
this tree properly. We need another solution to be able to pickle this
correctly. For now we revert the change to avoid generating invalid TASTy.