Port tests form scala2-library-tasty-tests to normal tests
Split BootstrappedStdLibTASYyTest into the part that tests the TASTy
inspector and the part that tests recompilation.

We also remove `scala2-library-tasty-tests` as it is now empty and will
not serve any purpose anymore.

This test also discovered a bug in TASTYRun with the path separators of
`JarArchive`s in Windows. The "/" is used in the JAR paths but the "\"
is used for OS paths. This is now fixed.