Allow types in given definitions to be infix types
A type implemented in a given definition can now be an infix type, without
enclosing parens being necessary.

By contrast, it cannot anymore be a refined type. Refined types have to be
enclosed in parens. This second point aligns the dotty parser with the
published syntax and the scala meta parser.