Infer more self types automatically
Clean up the logic how we infer self types, and add a new clause:

> If we have an externally extensible class that itself does not have a
declared self types itself and also not in any of its base classes,
assume {cap} as the self type. Previously we would install a capture
set but then check after the fact that that capture set is indeed {cap}.
So it's less verbose to just assume that from the start.