fix: false-positive closureDef match
Before this commit, closureDef matching rule was not strict enough to
filter out false-positive cases, which caused
tests/init-global/neg/return2.scala to fail.

It matched a plain block with ValDef at the begining and Closure at
the end.

As we want to match `closureDef` against a closure or a synthetic closure generated from
eta-expansion, we should match against

1. a simple closure
2. a closure inside nested blocks without statement, which is generated from eta-expansion on method without
   default arguments
3. a closure inside nested blocks with statements consisting only of ValDef of
   default arguments, which is generated from eta-expansion on
   method with default arguments.

This commit adds the condition for 3.