Error instead of StaleSymbol crash for certain cyclic macro dependencies
If a suspended macro refers back to a symbol in a previously compiled package
object, report this as an error instead of crashing with a StaleSymbol exception.

There's unfortunately not a lot of info available at the point where the error
is raised, so the error message is a bit vague. But it's better than crashing.

[Cherry-picked e4e1c57168bac79df7c4affff38d408127e9caaa]