fix(#16610): warn Scaladoc on multiple cases
Before this commit, the compiler ignored Scaladoc comment on multiple
enum cases without warning.

This is partly expected because the case to which the doc is attached is ambiguous,
but we should at least warn users that the comment is ignored by compiler due to
ambiguity and they should take an action if they want the doc to be displayed.