Don't try default args when there are ambiguous implicits
Before this commit, the logic in `class Typer` / `def adapt1` / `def addImplicitArgs` was to try default arguments if *any* argument had an implicit search error that was not an ambiguous implicits error.

This commit changes the logic to try default arguments only if *all* arguments' implicit search errors are not ambiguous implicits error.

The rational is that if there is any ambiguous implicits error, it's not worth trying default arguments because this cannot solve the ambiguity. Instead, we report the first ambiguous implicits error directly.