Inline transparent implicit parameters when typing Unapply trees
We needed to delay the inlining of the transparent inline when typing
the unapply function application. We used the NoInline mode, but this
also stopped the inlining of the arguments of the unapply. To fix this
we target more precisely the inlining of the unapply method and not the
implicit arguments. To do this we detect the dummy argument that is used
type the unapply as an application, before it is transformed into a pattern.

Fixes #19623

[Cherry-picked d244a30b67ce6db9e929f3e7e9c0ef4051ac0e5c]