fix(#i18645): overload ext method body in braces didn't compile
`tryWithImplicitOnQualifier` returned without attempting `tryInsertImplicitOnQualifier`,
which prevented the compiler from finding overloaded method, because `hasInnerErrors`
in `ProtoTypes` picked up the inner error propagated back from outer error.

When compiling `x.pprint(()=> { 123 })` before this commit, in `hasInnerErrors`,
`t` contains two sub trees; `Block(stmt, expr)` and `stmt`. The span of the former matches
the span of `t`, but that of the latter does not, which caused `hasInnerErrors` to return `true`,
but the error from the latter is duplicated with the error of parent tree `t`.

[Cherry-picked 5cf670741afaa6eb9bc670f6693fbb82fc5b4c01]