More careful type variable instance improvements
The previous code tried to recursively apply the current instance
of FullyDefinedAccumulator to the prospective instance type. This
can have unforeseen side-effects, as i19637 shows. We now are more
conservative: We check that the prospective instance type is already
fully defined without the possibility to instantiate more type variables.
This still passes the test cases that type variable improvement solves
and avoids the problem with #19637.

Fixes #19637