Do not enable experimental by default in nightly and snapshot versions
This scheme was devised as a simple way to enable experimental on non-stable
released. This made it impossible to publish a library containing experimental
features for a non-experimental version of the compiler. We had `-Yno-experimental`
to test the expected behavior on stable releases. Later we added `-experimental`
which add `@experimental` to all top-level classes. On a stable release, this
has the same effect as enabling experimental features, but it also allows
publishing this library (users will need to also be experimental to use it).

We can use `-experimental` on nightly and snapshot releases to achieve the
same behavior as before. By doing this the `-Yno-experimental` flag is no
longer needed and can be removed.

We will need to use `-experimental` in all our tests to enable experimental
by default. We can add this compiler flag to the default settings. We
also need a few tests that do not use `-experimental` to test the behavior
of `@experimental`. These can be added in `tests/*-no-experimental/`. Not
all tests have `-experimental` enabled, this will be added on a need basis.

With this we detected one bug, the runtime staging compile crashes if the
code uses experimental features. To fix this we enable `-experimental` by
default in this compiler. Note that the experimental-ness checks will have
been done by the compiler that compiled the quotes.