Drop FreeSeqFactory from stdlib-cc
FreeSeqFactory was a construction to demonstrate type safety for certain iterableFactory.from calls
where we rely in the fact that for all Seqs iterableFactory has an eager implementation of from.

While that shows that we _can_ make it typesafe, it does not work at runtime as a drop-in replacement
for stdlib since of course stdlib does not have a FreeSeqFactory.

This commit drop FreeSeqFactory and adds three unsafeAssumePure calls instead, with explanations.

Fixes #19745