Simplify class instance initialization check
Previously we had a rule that checks whether a warm object can be promoted by checking
its member methods and the corresponding return values recursively. The rule is complex and produce
error messages that are difficult to understand.

Initially, the rule was intended for the following use case:

    object C:
      case class Data(value: Int)
      val a = Data(3)
      val ref = foo(a) // safe to leak a

With the introduction of the global object checker, we do not check initialization of global objects
with the class instance checker anymore. Consequently, we can simplify the rules of class instance checker.