Allow poly-kinded GADT vars to infer constraints
Allowing the types to flow through in thirdTryNamed requires a tweak to
baseType.  Given "class C[T <: Int]" (from tests/pos-macros/i10864a) the
basetype of a TypeRef of just "C", with the Any class symbol, shouldn't
return Any because "C" is not a proper/value type.  That is, in
baseType, we need to short-circuit TypeProxy.superType (i.e.
underlying), which is returns resultType for a HKLambda.