Account for match types reducing to `ErrorType`s in TypeComparer
A match type has no matching cases is reduced to an `ErrorType` (different from throwing a TypeError), in distinction to the other cases for failed match type reduction which return `NoType`.

Producing to an ErrorType is usually accompanied by reporting an error, but does appear to be the case when attempting normalization from within the TypeComparer.