Update check-files and error annotations
The diff in neg/10349.scala is quite interesting.
With a few intermediate values:
```scala
type First[X] = X match
 case Map[_, v] => First[Option[v]]

def first[X](x: X): First[X] = x match
 case x: Map[k, v] =>
 val hdOpt: Option[v] = x.values.headOption
 first(hdOpt): First[Option[v]] // error only before changes
```
This now type-checks but will fail at runtime because of the in-exhaustivity of the match expression.
Perhaps we should add some additional condition in `isMatchTypeShaped` to account for this, or at least emit a warning ?