Allow types in given definitions to be infix types
A type implemented in a given definition can now be an infix type, without
enclosing parens being necessary.

By contrast, it cannot anymore be a refined type. Refined types have to be
enclosed in parens. This second point aligns the dotty parser with the
published syntax and the scala meta parser.

# Conflicts:
#	tests/pos/typeclasses-this.scala

[Cherry-picked ef71dcb45a0f31b72c5fe05fc48764865e1cea8e]