Strengthen MiMa checks
Using the latest patch release of the current minor version may lead to
some breakages that are not properly checked/tagged in the MiMa filers.
For example, if we accidentally introduce a breaking change in 3.M.1 but
check against 3.M.2, we would lose this braking change before the next
minor release. If we would check against 3.M.0, we would catch this.

In general we need to track all the changes that have happened in the
current minor version. There should not be any breaking changes unless
the PR needs a minor release.