Implement match type amendment: extractors follow aliases and singletons
This implements the change proposed in
https://github.com/scala/improvement-proposals/pull/84.

The added pos test case presents motivating examples, the added neg test cases
demonstrate that errors are correctly reported when cycles are present. The
potential for cycle is no worse than with the existing extraction logic as
demonstrated by the existing test in `tests/neg/mt-deskolemize.scala`.

[Cherry-picked 1a235c6719f56e1597241dc38eeda49087b323e8]