Rewrite reaches tests to fix tasty-cc nightly failures
The reaches.scala pos test failed the nightly with tasty cc since one line violated
the new condition for reaches capabilities. The problem was not discoveed in normal
testing since it used the `List` class, which is capture checked only in the nightly.
We now make the relevant bits of `List` explicit in the test itself, so that it is
always capture checked. The failing line got moved to neg.