Retain default parameters with `export`
Default parameters need to have the `HasDefault` flag set to work properly, it
seems that without this flag they were still selected but only under join
compilation.

While we're at it, we define a full set of flags that might be exported. Note
that Given/Implicit/Erased were already set by `tpd.DefDef` and Inlined doesn't
seem to make a difference in practice since the body of the exported inline def
is manually constructed and doesn't contain proxies either way.