Stabilise returned completions by improving deduplication + extra completions for constructors (#19976)
This PR doesn't address all issues.
In the future we need to get rid of
https://github.com/scala/scala3/compare/main...rochala:improved-deduplication-and-constructors-search?expand=1#diff-035851592480495dfdb20da6b615ec7dd77b3db70cda46aba56230d8cd690773R157-R167
as it is not working as intended. The future PR of shortened type
printer refactor includes a proper way to split extension params.
`CompletionValue.Interpolator` should also be removed, and instead we
should return workspace / completions as it is now hard to sort those
completions.

Next refactor is reusing completion affix for other kinds of completions
such as case completions, so prefix / suffix is handled in single place.

This PR will unblock fuzzy search in the compiler because of stabilizing
returned completions.