Unify export target validation.
Instead of having separate checks for members and classes/modules, we
unify them. This gives simpler code and will make it easier to allow
nested objects.

Additionally: We report a proper error when attempting to export a JS
native member (instead of crashing the compiler).

Forward port of the upstream commit
https://github.com/scala-js/scala-js/commit/14de17dc632d5a4983bdcce18f79c09b3ceb61bc