DropSkolemMap: simplify logic
No need to save the value of `refersToSkolem`: if it's true before we enter
`NamedType` it will be true after and `dropSkolem` will return `NoType`.

The previous logic could still be useful if we want to give more easily
actionable error messages in the future by only keeping in the type the skolems
we couldn't remove.

[Cherry-picked a1930c4ca38673885a4ebc2ce95689e9e65d08be]